+++ qwpython-src/qwsv/net.h
@@ -39,7 +39,7 @@ extern	int		net_socket;
 void		NET_Init (int port);
 void		NET_Shutdown (void);
 qboolean	NET_GetPacket (void);
-void		NET_SendPacket (int length, void *data, netadr_t to);
+void		NET_SendPacket(size_t length, void *data, netadr_t to);
 
 qboolean	NET_CompareAdr (netadr_t a, netadr_t b);
 qboolean	NET_CompareBaseAdr (netadr_t a, netadr_t b);
